Add test for setCellFlags
Nearly all the library linkages were wrong. The rule is to link the library to the target where it is used. PRIVATE if it is in cpp code only, INTERFACE if it is in header code only, and PUBLIC if it is in both. Fixing this was required to get the test to link properly.

strings.h was invalid as a header. It could have been fixed by marking all the definitions inline, but I chose to put them in a new file instead.
