+++ CorsixTH/CMakeLists.txt
@@ -121,7 +121,7 @@ else()
       message("Make sure the path is correctly defined or set the environment variable SDLDIR to the correct location")
     endif()
     # No need to specify sdl2main separately, the FindSDL.cmake file will take care of that. If not we get an error about it
-    target_link_libraries(CorsixTH PRIVATE ${SDL_LIBRARY})
+    target_link_libraries(CorsixTH_lib PUBLIC ${SDL_LIBRARY})
     message("  SDL2 found")
   else()
     message(FATAL_ERROR "Error: SDL2 library not found, it is required to build. Make sure the path is correctly defined or set the environment variable SDLDIR to the correct location")
@@ -132,11 +132,10 @@ endif()
 if(VCPKG_TARGET_TRIPLET)
   find_package(Lua CONFIG REQUIRED)
   target_link_libraries(CorsixTH_lib PUBLIC lua)
-  target_link_libraries(CorsixTH PRIVATE lua)
 else()
   find_package(Lua REQUIRED)
   if(Lua_FOUND OR LUA_FOUND)
-    target_link_libraries(CorsixTH PUBLIC ${LUA_LIBRARY})
+    target_link_libraries(CorsixTH_lib PUBLIC ${LUA_LIBRARY})
     include_directories(${LUA_INCLUDE_DIR})
     # Special link flags needed on macOS 64bit, according to: http://luajit.org/install.html
     # If these are not specified, luaL_newstate() returns NULL and we get this:
@@ -167,13 +166,13 @@ if(CORSIX_TH_USE_SDL_MIXER)
   if(VCPKG_TARGET_TRIPLET)
     find_package(SDL2_mixer CONFIG REQUIRED)
     target_link_libraries(
-      CorsixTH
-      PRIVATE
+      CorsixTH_lib
+      PUBLIC
         $<IF:$<TARGET_EXISTS:SDL2_mixer::SDL2_mixer>,SDL2_mixer::SDL2_mixer,SDL2_mixer::SDL2_mixer-static>)
   else()
     find_package(SDL2_mixer REQUIRED)
     if(SDLMIXER_FOUND)
-      target_link_libraries(CorsixTH PRIVATE ${SDLMIXER_LIBRARY})
+      target_link_libraries(CorsixTH_lib PUBLIC ${SDLMIXER_LIBRARY})
       include_directories(${SDLMIXER_INCLUDE_DIR})
       message("  SDL_mixer found")
     else()
@@ -186,7 +185,7 @@ endif()
 if(CORSIX_TH_USE_FFMPEG)
   find_package(FFmpeg COMPONENTS AVFORMAT AVCODEC AVUTIL SWSCALE SWRESAMPLE REQUIRED)
   if(FFMPEG_FOUND)
-    target_link_libraries(CorsixTH PRIVATE ${FFMPEG_LIBRARIES})
+    target_link_libraries(CorsixTH_lib PUBLIC ${FFMPEG_LIBRARIES})
     include_directories(${FFMPEG_INCLUDE_DIRS})
     if(APPLE)
       target_link_libraries(CorsixTH PRIVATE libz.dylib)
@@ -203,15 +202,15 @@ endif()
 if(CORSIX_TH_USE_FREETYPE2)
   if(VCPKG_TARGET_TRIPLET)
     find_package(Freetype REQUIRED)
-    target_link_libraries(CorsixTH PRIVATE Freetype::Freetype)
+    target_link_libraries(CorsixTH_lib PUBLIC Freetype::Freetype)
   else()
     find_package(Freetype REQUIRED)
     if(FREETYPE_FOUND)
-      target_link_libraries(CorsixTH PRIVATE ${FREETYPE_LIBRARIES})
+      target_link_libraries(CorsixTH_lib PUBLIC ${FREETYPE_LIBRARIES})
       include_directories(${FREETYPE_INCLUDE_DIRS})
       if(APPLE)
-        target_link_libraries(CorsixTH PRIVATE libz.dylib)
-        target_link_libraries(CorsixTH PRIVATE libbz2.dylib)
+        target_link_libraries(CorsixTH_lib PRIVATE libz.dylib)
+        target_link_libraries(CorsixTH_lib PRIVATE libbz2.dylib)
       endif()
       message("  FreeType2 found")
     else()
@@ -223,7 +222,7 @@ endif()
 if(WITH_UPDATE_CHECK)
   find_package(CURL REQUIRED COMPONENTS HTTPS SSL)
   if(CURL_FOUND)
-    target_link_libraries(CorsixTH PRIVATE CURL::libcurl)
+    target_link_libraries(CorsixTH_lib PRIVATE CURL::libcurl)
     message("  Curl found")
   else()
     message(FATAL_ERROR "Error: curl not found")
@@ -233,10 +232,10 @@ endif()
 if(CORSIX_TH_LINK_LUA_MODULES)
   message("Linking lua modules")
   find_package(unofficial-luafilesystem CONFIG REQUIRED)
-  target_link_libraries(CorsixTH PRIVATE unofficial::luafilesystem::lfs)
+  target_link_libraries(CorsixTH_lib PRIVATE unofficial::luafilesystem::lfs)
 
   find_package(unofficial-lpeg CONFIG REQUIRED)
-  target_link_libraries(CorsixTH PRIVATE unofficial::lpeg::lpeg)
+  target_link_libraries(CorsixTH_lib PRIVATE unofficial::lpeg::lpeg)
 
   add_custom_command(TARGET CorsixTH POST_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy
