+++ CorsixTH/CppTest/test_th_strings.cpp
@@ -1,9 +1,9 @@
-#include <string>
-
-#include "../Src/th_strings.h"
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers.hpp>
 #include <catch2/matchers/catch_matchers_string.hpp>
+#include <string>
+
+#include "../Src/th_strings.h"
 
 TEST_CASE("skip whitespace", "[skip_utf8_whitespace]") {
   const char* str = " \t \n  1234";
