+++ CorsixTH/Src/CMakeLists.txt
@@ -31,6 +31,7 @@ target_sources(CorsixTH_lib
     ${CMAKE_CURRENT_SOURCE_DIR}/th_movie.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/th_pathfind.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/th_sound.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/th_strings.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/xmi2mid.cpp
 )
 
@@ -57,6 +58,7 @@ target_sources(CorsixTH_lib
     ${CMAKE_CURRENT_SOURCE_DIR}/th_movie.h
     ${CMAKE_CURRENT_SOURCE_DIR}/th_pathfind.h
     ${CMAKE_CURRENT_SOURCE_DIR}/th_sound.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/th_strings.h
     ${CMAKE_CURRENT_SOURCE_DIR}/xmi2mid.h
 
     ${CMAKE_BINARY_DIR}/CorsixTH/Src/config.h
