+++ CorsixTH/Src/lua.hpp
@@ -25,10 +25,10 @@ SOFTWARE.
 
 extern "C" {
 // IWYU pragma: begin_exports
+#include <lauxlib.h>
 #include <lua.h>
 #include <luaconf.h>
 #include <lualib.h>
-#include <lauxlib.h>
 // IWYU pragma: end_exports
 }
